pyproject.toml
[tool.poetry]
name = "prompt-engineering"
version = "0.1.0"
description = ""
authors = ["CommonGibbon <xenoproboscizoid@gmail.com>"]
readme = "README.md"
packages = [{include = "prompt_engineering", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
hydra-core = "^1.3.2"
mlflow = "^3.5.1"
jupyterlab = "^4.4.10"
pandas = "^2.3.3"
numpy = "^2.3.4"
ipykernel = "^7.0.1"
pillow = "^12.0.0"
openai = "^2.6.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"